% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mod.info}
\alias{mod.info}
\title{Model Information and Performance Evaluation}
\usage{
mod.info(model, data, verbose = TRUE, save.acc = FALSE)
}
\arguments{
\item{model}{A trained classification model. It could be an object of class `nnet::multinom`
or `MASS::polr`.}

\item{data}{A data frame containing the predictors and the true class labels in a column
named `class`.}

\item{verbose}{Logical; if `TRUE` (default), prints out the model's accuracy, pseudo R-squared,
and coefficients.}

\item{save.acc}{Logical; if `TRUE`, saves the model's accuracy as a global variable `Acc.print`
for further use. Default is `FALSE`.}
}
\value{
A list containing:
\item{class.table}{The confusion matrix of actual vs. predicted classes.}
\item{Accuracy}{The computed accuracy of the model as a percentage.}
\item{McFadden_R2}{The McFadden pseudo R-squared value of the model.}
}
\description{
This function computes and displays key performance metrics for a trained classification model,
including accuracy and McFadden's pseudo R-squared. It also generates a confusion matrix and
displays the model coefficients.
}
\details{
The function generates predictions for the entire dataset and compares them with the true class
labels. It calculates the confusion matrix, overall accuracy, and McFadden's pseudo R-squared
for evaluating the model's goodness of fit. For multinomial regression models (`nnet::multinom`
or `MASS::polr`), the log-likelihood of the null model (intercept-only) is compared to that of
the fitted model to compute the pseudo R-squared.

If `verbose` is set to `TRUE`, the function outputs:
\itemize{
  \item The confusion matrix.
  \item The accuracy of the model.
  \item McFadden's pseudo R-squared value.
  \item The model coefficients.
}
}
